Update run_tmux.sh
Checking a session with tmux has-session:

I replaced tmux ls / grep-o "$sessionName" with a cleaner and more correct command tmux has-session -t "$sessionName " 2>/dev/null. This allows you to correctly check whether a session named $sessionName exists, and avoid errors if tmux is not running or the session is not found.
Checking the configuration for root:

Added a check for the existence of the ~/.tmux/.tmux.template.conf configuration file before loading it with .. If there is no file, a message is displayed.
Checking the TERM color:

Replaced the condition [ $TERM == '*256color*'] with a more correct use of regular expressions using [["$TERM " = ~ 256color ]], which allows you to check whether the TERM variable contains the string 256color.
Fixed commands for running tmux:

I fixed the commands with tmux -2 and tmux -8. Now they are executed correctly when the session starts, and not after connecting, since these flags are needed in order for tmux to process colors correctly.
Displaying messages:

Added message output at each important step to make it clear what the script does (for example, when creating a new session or loading a configuration).

Уremoved the -8 flag: tmux does not support the -8 flag, so I removed it.
Support for 256 colors with -2: For 256 colors, use the -2 flag, which is enabled when the TERM variable is configured accordingly.
Default is 88 colors: When TERM is not configured for 256 colors, tmux defaults to 88 colors, and therefore no additional flags are required. Just start the session without the -2 flag to use the standard palette.